# Change the regex, increase performance increadibly :D