--- a/src/http.js
+++ b/src/http.js
@@ -64,12 +64,7 @@ export default function http(url, request = {}) {
   })
 }
 
-function shouldDownloadAsText(contentType) {
-  return /json/.test(contentType) ||
-         /xml/.test(contentType) ||
-         /yaml/.test(contentType) ||
-         /text/.test(contentType)
-}
+const shouldDownloadAsText = (contentType = "") => /json|xml|yaml|text/.test(contentType)
 
 // Serialize the response, returns a promise with headers and the body part of the hash
 export function serializeRes(oriRes, url, {loadSpec = false} = {}) {