--- a/test/http.js
+++ b/test/http.js
@@ -1,7 +1,10 @@
 import expect from 'expect'
 import xmock from 'xmock'
 import fetchMock from 'fetch-mock'
-import http, {serializeHeaders, mergeInQueryOrForm, encodeFormOrQuery, serializeRes} from '../src/http'
+import http, {
+  serializeHeaders, mergeInQueryOrForm, encodeFormOrQuery, serializeRes,
+  shouldDownloadAsText
+} from '../src/http'
 
 describe('http', () => {
   let xapp
@@ -301,6 +304,33 @@ describe('http', () => {
         expect(resSerialize.data).toBe(resSerialize.text)
         expect(resSerialize.data).toBe(body)
       }).then(fetchMock.restore)
+    })
+  })
+
+  describe('shouldDownloadAsText', () => {
+    it('should return true for json, xml, yaml, and text types', function() {
+      const types = [
+        "text/x-yaml", "application/xml", "text/xml", "application/json",
+        "text/plain"
+      ]
+
+      types.forEach(v => {
+        expect(`${v} ${shouldDownloadAsText(v)}`).toEqual(v + " true")
+      })
+    })
+
+    it('should return false for other common types', function() {
+      const types = [
+        "application/octet-stream", "application/x-binary"
+      ]
+
+      types.forEach(v => {
+        expect(`${v} ${shouldDownloadAsText(v)}`).toEqual(v + " false")
+      })
+    })
+
+    it('should fail gracefully when called with no parameters', function() {
+      expect(shouldDownloadAsText()).toEqual(false)
     })
   })
 })