first vs last
```
function shouldDownloadAsTextOrig(contentType) {
  return /json/.test(contentType) ||
        /xml/.test(contentType) ||
        /yaml/.test(contentType) ||
        /text/.test(contentType)
}
/*
Orig|json x 11,477,916 ops/sec ±0.75% (91 runs sampled)
Orig|xml x 7,497,474 ops/sec ±0.52% (97 runs sampled)
Orig|yaml x 5,051,156 ops/sec ±0.68% (94 runs sampled)
Orig|text x 4,224,534 ops/sec ±0.39% (96 runs sampled)
*/

function shouldDownloadAsTextOne(contentType = "") {
  return ["json", "xml", "yaml", "text"].some(str => contentType.includes(str))
}
/*
One|json x 6,393,355 ops/sec ±0.62% (94 runs sampled)
One|xml x 4,242,619 ops/sec ±0.89% (94 runs sampled)
One|yaml x 2,880,927 ops/sec ±0.35% (95 runs sampled)
One|text x 2,286,908 ops/sec ±0.61% (93 runs sampled)
*/

function shouldDownloadAsTextTwo(contentType = "") {
    return /json|xml|yaml|text/.test(contentType)
}
/*
Two|json x 19,513,133 ops/sec ±0.29% (92 runs sampled)
Two|xml x 24,767,680 ops/sec ±0.70% (93 runs sampled)
Two|yaml x 24,596,672 ops/sec ±0.31% (94 runs sampled)
Two|text x 24,863,942 ops/sec ±0.68% (88 runs sampled)
*/

const shouldDownloadAsTextThree = (contentType = "") => /json|xml|yaml|text/.test(contentType)
/*
three|json x 20,038,598 ops/sec ±0.31% (93 runs sampled)
three|xml x 24,801,552 ops/sec ±0.68% (92 runs sampled)
three|yaml x 25,187,765 ops/sec ±0.47% (94 runs sampled)
three|text x 25,252,579 ops/sec ±0.67% (96 runs sampled)
*/
```